# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Customer Controller
  description: ""
  termsOfService: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Joel Silva
    email: joeltadeu@gmail.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
  - url: http://localhost:9081
    description: Generated server url
...
paths:
  /v1/customers/{id}:
    get:
      tags:
        - Customer Controller
      summary: Retrieve a customer by id
      description: This endpoint is responsible to retrieve the customer data by id
      operationId: findCustomerById
      parameters:
        - name: id
          in: path
          description: Id of the customer to be searched
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "404":
          description: Customer not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to retrieve
                    a customer does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Customer with id 10 was not found
                    date: 2022-10-28T21:39:54.603263862
        "200":
          description: Return customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        "500":
          description: An unexpected error occurred during retrieve the customer
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to retrieve a customer
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
    put:
      tags:
        - Customer Controller
      summary: Update a customer
      description: This endpoint is responsible to update the customer data by id
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          description: Id of the customer to be updated
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
        required: true
      responses:
        "400":
          description: Customer request is invalid
          content:
            application/json:
              examples:
                Bad request example:
                  description: A bad request response example when trying to update
                    a customer
                  value: |
                    {
                         "code": 400,
                         "status": "Bad Request",
                         "description": "Validation Exception",
                         "date": "2022-11-25T20:14:04.078313400",
                         "attributes": [
                             {
                                 "attribute": "firstName",
                                 "errors": [
                                     "First name cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "lastName",
                                 "errors": [
                                     Last name cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "birthday",
                                 "errors": [
                                     "Birthday cannot be null"
                                 ]
                             }
                         ]
                    }
        "404":
          description: Customer not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to update
                    a customer does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Customer with id 10 was not found
                    date: 2022-10-28T21:39:54.603263862
        "200":
          description: Customer updated
          content:
            application/json: {}
        "500":
          description: An unexpected error occurred during update the customer
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to update a customer
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
    delete:
      tags:
        - Customer Controller
      summary: Delete the customer by id
      description: This endpoint is responsible to delete the customer by id
      operationId: deleteCustomer
      parameters:
        - name: id
          in: path
          description: Id of the customer to be deleted
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "404":
          description: Customer not found
          content:
            application/json:
              examples:
                Not found example:
                  description: A not found response example when trying to delete
                    a customer does not exist
                  value:
                    code: 404
                    status: Not Found
                    description: Customer with id 10 was not found
                    date: 2022-10-28T21:39:54.603263862
        "204":
          description: Customer deleted
          content:
            application/json: {}
        "500":
          description: An unexpected error occurred during delete the customer
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to delete a customer
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
...
  /v1/customers:
    get:
      tags:
        - Customer Controller
      summary: Retrieve a list of customers
      description: This endpoint is responsible to retrieve a list of customers based
        on the applied filters
      operationId: listAllCustomers
      parameters:
        - name: filter
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/CustomerFilter'
      responses:
        "500":
          description: An unexpected error occurred during retrieve the customers
            list
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to retrieve a customers list
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "200":
          description: Return customers list
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCustomerResponse'
    post:
      tags:
        - Customer Controller
      summary: Create a customer
      description: This endpoint is responsible to create a new customer
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
        required: true
      responses:
        "400":
          description: Customer request is invalid
          content:
            application/json:
              examples:
                Bad request example:
                  description: A bad request response example when trying to create
                    a customer
                  value: |
                    {
                         "code": 400,
                         "status": "Bad Request",
                         "description": "Validation Exception",
                         "date": "2022-11-25T20:14:04.078313400",
                         "attributes": [
                             {
                                 "attribute": "firstName",
                                 "errors": [
                                     "First name cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "lastName",
                                 "errors": [
                                     Last name cannot be null"
                                 ]
                             },
                             {
                                 "attribute": "birthday",
                                 "errors": [
                                     "Birthday cannot be null"
                                 ]
                             }
                         ]
                    }
        "500":
          description: An unexpected error occurred during creating the customer
          content:
            application/json:
              examples:
                Internal server error example:
                  description: A internal server error response example when trying
                    to create a customer
                  value:
                    code: 500
                    status: Internal Server Error
                    description: An error occurred while processing your request
                    date: 2022-10-28T21:39:54.603263862
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
...
components:
  schemas:
    CustomerRequest:
      required:
        - birthday
        - firstName
        - lastName
      type: object
      properties:
        firstName:
          type: string
          description: Customer first name
          example: Damian
        lastName:
          type: string
          description: Customer last name
          example: O’Brien
        birthday:
          type: string
          description: Customer birthday
          format: date
          example: 1981-04-21
...
    AddressResponse:
      type: object
      properties:
        id:
          type: integer
          description: Address Id
          format: int64
          example: 13
        street:
          type: string
          description: Address street
          example: Coosan Road
        complement:
          type: string
          description: Address complement
          example: "Apt. 104, Block 4"
        eircode:
          type: string
          description: Address eircode
          example: N46RP87
        city:
          type: string
          description: Address city
          example: Athlone
        county:
          type: string
          description: Address county
          example: Co. Westmeath
        country:
          type: string
          description: Address country
          example: Ireland
    CustomerResponse:
      type: object
      properties:
        id:
          type: integer
          description: Customer Id
          format: int64
          example: 13
        firstName:
          type: string
          description: Customer first name
          example: Damian
        lastName:
          type: string
          description: Customer last name
          example: O’Brien
        birthday:
          type: string
          description: Customer birthday
          format: date
          example: 1981-04-21
        addresses:
          type: array
          description: Customer addresses list
          items:
            $ref: '#/components/schemas/AddressResponse'
        emails:
          type: array
          description: Customer emails list
          items:
            $ref: '#/components/schemas/EmailResponse'
        documents:
          type: array
          description: Customer documents list
          items:
            $ref: '#/components/schemas/DocumentResponse'
    DocumentResponse:
      required:
        - documentNumber
      type: object
      properties:
        id:
          type: integer
          description: Document Id
          format: int64
          example: 13
        type:
          type: string
          description: Document type
          example: PASSPORT
          enum:
            - DRIVER_LICENSE
            - PASSPORT
            - PPS
        documentNumber:
          type: string
          description: Document number
          example: 187263IN
    EmailResponse:
      type: object
      properties:
        id:
          type: integer
          description: Email Id
          format: int64
          example: 13
        type:
          type: string
          description: Email type
          example: PERSONAL
          enum:
            - WORK
            - PERSONAL
        email:
          type: string
          description: Email
          example: damian.obrien@gmail.com
    CustomerFilter:
      required:
        - pageNumber
        - pageSize
      type: object
      properties:
        pageNumber:
          type: integer
          description: Zero-base page index (0..N)
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: Number of records per page
          format: int32
          example: 10
          default: 10
        sortDirection:
          type: string
          description: Sort direction
          example: ASC|DESC
          enum:
            - ASC
            - DESC
        sortBy:
          type: string
          description: Sort column name
          example: firstName
        firstName:
          type: string
          description: First name to be searched
          example: Damian
        lastName:
          type: string
          description: Last name to be searched
          example: O’Brien
    PageCustomerResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
...